packages/aicp-core/python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aicp-core"
version = "0.1.0"
description = "AICP: AI Communication Protocol"
requires-python = ">=3.10"
dependencies = [
    "msgpack>=1.0.7",
    "cryptography>=41.0.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "pylint>=2.17.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=aicp --cov-report=html --cov-report=term"
